refactor: migrate more to xstate
