+++ app/consts.ts
@@ -9,8 +9,3 @@ export const WHITE = 'W'
 export const BLACK = 'B'
 export const WHITE_CANDIDATE = 'WC'
 export const BLACK_CANDIDATE = 'BC'
-
-// State
-export const IDLE = 'IDLE'
-export const PLAYING = 'PLAYING'
-export const ENDED = 'ENDED'
