+++ app/store/__tests__/selector.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
 
-import { computeScore } from '../compute-score'
-import { BLACK, BLACK_CANDIDATE, WHITE, WHITE_CANDIDATE } from '../consts'
+import { BLACK, BLACK_CANDIDATE, WHITE, WHITE_CANDIDATE } from '~/consts'
+import { computeScore } from '~/lib/compute-score'
 
 describe('selector.ts', () => {
   describe('computeScore', () => {
@@ -10,7 +10,7 @@ describe('selector.ts', () => {
         computeScore([
           [BLACK, WHITE],
           [WHITE, BLACK],
-        ]),
+        ] as any),
       ).toEqual({ white: 2, black: 2 })
     })
 
@@ -19,7 +19,7 @@ describe('selector.ts', () => {
         computeScore([
           [BLACK, WHITE, WHITE_CANDIDATE, BLACK_CANDIDATE],
           [WHITE, BLACK, null, null],
-        ]),
+        ] as any),
       ).toEqual({ white: 2, black: 2 })
     })
 
@@ -28,7 +28,7 @@ describe('selector.ts', () => {
         computeScore([
           [BLACK, WHITE_CANDIDATE, BLACK_CANDIDATE],
           [BLACK, null, null],
-        ]),
+        ] as any),
       ).toEqual({ white: 0, black: 2 })
     })
   })
