--- app/store/utils.ts
@@ -1,1 +1,0 @@
-export const createNull = (): null => null
