+++ app/types/board.ts
