--- tsconfig.app.json
@@ -1,37 +1,0 @@
-{
-  "compilerOptions": {
-    "composite": true,
-    "jsx": "react-jsx",
-    "lib": [
-      "dom",
-      "esnext"
-    ],
-    "baseUrl": ".",
-    "module": "es2022",
-    "moduleResolution": "Bundler",
-    "paths": {
-      "~/*": [
-        "./app/*"
-      ]
-    },
-    "resolveJsonModule": true,
-    "types": [
-      "vite/client",
-      "node",
-      "unplugin-icons/types/react"
-    ],
-    "allowJs": true,
-    "strict": true,
-    "strictNullChecks": true,
-    "noUnusedLocals": true,
-    "downlevelIteration": true,
-    "noEmit": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true
-  },
-  "include": [
-    "./app/**/*",
-    "app.config.ts",
-    "vitest.config.ts"
-  ]
-}
